% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_data.R
\name{eval_data}
\alias{eval_data}
\title{Title}
\usage{
eval_data(pack, data_txt = NULL, envir = NULL, render = TRUE)
}
\arguments{
\item{pack}{}

\item{data_txt}{}

\item{envir}{}

\item{render}{}
}
\description{
Title
}
