% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_post.R
\name{eval_post}
\alias{eval_post}
\title{Title}
\usage{
eval_post(pack, post_txt, samples, envir = NULL, render = TRUE)
}
\arguments{
\item{pack}{}

\item{post_txt}{}

\item{samples}{}

\item{envir}{}

\item{render}{}
}
\description{
Title
}
