% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_samples_and_post.R
\name{get_samples_and_post}
\alias{get_samples_and_post}
\title{Title}
\usage{
get_samples_and_post(model, variable.names, iterations, progress_iterations,
  pack, post_txt, post_envir, render, data_html, cur_trace, session = session)
}
\arguments{
\item{model}{}

\item{variable.names}{}

\item{iterations}{}

\item{progress_iterations}{}

\item{pack}{}

\item{post_txt}{}

\item{post_envir}{}

\item{render}{}

\item{data_html}{}

\item{cur_trace}{}

\item{session}{}
}
\description{
Title
}
