% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jags_samples.R
\name{jags_samples}
\alias{jags_samples}
\title{Title}
\usage{
flexjags::jags_samples(pack, model, data_html, post_txt, result, iterations,
  progress_iterations, n_chain, cur_trace, eliminate_data, render, session)
}
\arguments{
\item{pack}{}

\item{model}{}

\item{data_html}{}

\item{post_txt}{}

\item{result}{}

\item{iterations}{}

\item{progress_iterations}{}

\item{n_chain}{}

\item{cur_trace}{}

\item{eliminate_data}{}

\item{render}{}

\item{session}{}
}
\description{
Title
}
