% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_packs.R
\name{get_packs}
\alias{get_packs}
\title{Title}
\usage{
get_packs(check = TRUE, stop_on_error = FALSE,
  pack_dir = system.file("Rmd", "example_packs", package = "flexjags"))
}
\arguments{
\item{check}{}

\item{stop_on_error}{}

\item{pack_dir}{}
}
\description{
Title
}
